+++ src/index.ts
@@ -1,7 +1,7 @@
 import { Theme } from './types';
 import createTheme from './theme';
 import * as colors from './colors';
-import createPalette from './palette';
+import createPalette from './createPalette';
 import shadows from './shadows';
 import spacing from './spacing';
 import transitions from './transitions';