+++ src/theme.ts
@@ -1,4 +1,4 @@
-import createPalette from './palette';
+import createPalette from './createPalette';
 import createTypography from './createTypography';
 import shadows from './shadows';
 import transitions from './transitions';
@@ -12,7 +12,7 @@ import breakpoints from './breakpoints';
  * @param options 主题配置
  */
 export default function createTheme(options: ThemeOptions = {}): Theme {
-  const palette = options.palette || createPalette();
+  const palette = createPalette(options.palette);
   const typography = createTypography(options.typography);
 
   return {